15MAY2017
As a ancient promise. I should start to write something every day. Not for English skill, not for bloging, not for someone or some boring ideas, but to see how my sttitude to world changes along time. It is a SECRET dairy so I do not want to show it to any gentlemen.

Yunlai said 离开KAUST将再也找不到对学生如此好的学校. Is KAUST such good? It just gave is a peaceful place to do research. I am so greedy that I wanna get both money and famousity. Terrafied but excited to get an chance to argue with my home department. Cannot imagine if I lost.